kiro-book/requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.5
Mako==1.3.2
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.6
WTForms==3.1.2
zope.event==5.0
zope.interface==6.2
